Rename streamingInitialRoundIterators to streamingGroundingIterators and allow instantiation of streamingCacheIterators with typed StreamingTermStores. (#304)
Rename streamingInitialRoundIterators to streamingGroundingIterators and allow instantiation of streamingCacheIterators with typed StreamingTermStores.
