+++ psl-core/src/main/java/org/linqs/psl/reasoner/dcd/term/DCDStreamingTermStore.java
@@ -65,7 +65,7 @@ public class DCDStreamingTermStore extends StreamingTermStore<DCDObjectiveTerm> 
 
     @Override
     protected StreamingIterator<DCDObjectiveTerm> getGroundingIterator() {
-        return new DCDStreamingInitialRoundIterator(
+        return new DCDStreamingGroundingIterator(
                 this, rules, atomManager, termGenerator,
                 termCache, termPool, termBuffer, volatileBuffer, pageSize, numPages);
     }
